refactor: convert firebase module to monorepo library
