+++ libs/firebase/src/constants/index.ts
