+++ libs/firebase/src/index.ts
@@ -1,0 +1,3 @@
+export * from './firebase.module';
+export * from './services';
+export * from './constants';
