+++ libs/firebase/src/interfaces/firebase-async-options.interface.ts
