+++ libs/firebase/src/interfaces/firebase-options-factory.interface.ts
