+++ libs/firebase/src/interfaces/firebase-options.interface.ts
