+++ libs/firebase/src/interfaces/index.ts
