+++ libs/firebase/src/services/firebase.service.ts
@@ -2,8 +2,8 @@ import { Inject, Injectable } from '@nestjs/common';
 import { initializeApp, App } from 'firebase-admin/app';
 import { getAuth, Auth } from 'firebase-admin/auth';
 import { getMessaging, Messaging } from 'firebase-admin/messaging';
-import { FIREBASE_OPTIONS } from '@modules/firebase/constants';
-import { FirebaseOptions } from '@modules/firebase/interfaces';
+import { FIREBASE_OPTIONS } from '../constants';
+import { FirebaseOptions } from '../interfaces';
 
 @Injectable()
 export class FirebaseService {
