--- src/modules/firebase/index.ts
@@ -1,1 +1,0 @@
-export * from './firebase.module';
